requirements.txt
alembic==1.4.2
click==7.1.2
Flask==1.1.2
Flask-Migrate==2.5.2
Flask-SQLAlchemy==2.4.1
itsdangerous==1.1.0
Jinja2==2.11.2
Mako==1.1.3
MarkupSafe==1.1.1
psycopg2==2.8.4
python-dateutil==2.8.1
python-editor==1.0.4
PyYAML==5.3.1
six==1.15.0
SQLAlchemy==1.3.18
Werkzeug==1.0.1
